Coach Health gets its own page, and its ratio stops flattering itself
Two things from the first live reading of the panel on production.

1. ITS OWN PLACE IN THE NAV. It was a card wedged into Overview, below six
   business stat tiles and above the scaling radar — a page about whether the
   coach is behaving, competing for room with MRR. It is now /coach-health with
   its own sidebar link, and Overview no longer renders it.

2. THE DENOMINATOR WAS THE WRONG SCOPE. The panel showed:

       Distance question answered without the distance
       0 of 1 matching request · 1 before the fix

   which reads as "one client asked since the fix and got it right". Nobody had
   asked since the fix. That single request was the pre-fix one. The numerator
   excluded pre-fix hits; the denominator counted every matching turn in the
   window including pre-fix ones — so the two halves of the ratio sat on
   opposite sides of the merge.

   An untested rule therefore displayed as a verified one. That is the flattering
   direction, and on a page whose only job is to be believed it is the one that
   matters. Both halves are now scoped to the same side, and the pre-fix pair is
   reported separately:

       0 of 0 requests since the fix · 1/1 before the fix

   which says what is true: this rule has not been exercised since it was fixed.

Graded over a synthetic window straddling a merge instant: one turn either side,
each landing on exactly one side, never summed.
